refactor: Remove unnecessary read
The read of aToken balance is unnecessary as the zkAToken already can be used to handle this accounting. The only case where `underlyingAmount > aTokenBalance` is if the rollup will transfer more zkATokens than it has ever minted.
Added a fuzzing test for the rounding case to showcase that the values will be exact, as the 1 that is added by rounding, will be truncated in the division.