fix: Simplify rounding to always round down
The Aave protocol perform rounding both up and down, which can make it a bit odd in behaviour as it might not always seem consistent. To enhance consistency of the expected behaviour, we now round down always. Meaning that we will underestimate by up to 1 wei when compared to Aave for each action.